DS cascade delete tags

+++ TabloidCLI/Repositories/TagRepository.cs
@@ -83,7 +83,7 @@ namespace TabloidCLI
                 conn.Open();
                 using (SqlCommand cmd =conn.CreateCommand())
                 {
-                    cmd.CommandText = @"DELETE FROM Tag WHERE Id = @id";
+                    cmd.CommandText = @"DELETE FROM AuthorTag WHERE TagId = @id; DELETE FROM BlogTag WHERE TagId = @id; DELETE FROM PostTag WHERE TagId = @id; DELETE FROM Tag WHERE Id = @id";
                     cmd.Parameters.AddWithValue("@id", id);
                     cmd.ExecuteNonQuery();
                 }
